get rid of the stale ExportBuilder class and use 'organizer' instead of 'export' for the organizer that's passed in on initialize
